Handle traffic to legacy urls (#161)
* Use standard naming for vercel config file

For backwards compatibility reasons `now.json` is still supported. The
standard, documented convention for this configuration file is now
`vercel.json`.
https://vercel.com/docs/platform/frequently-asked-questions#conflicting-configuration-files

* Add custom 404 page that links to legacy app

There are a lot of routes supported by the current app that may not be
covered by egghead-next. This 404 page acts as a catch-all prompting
users to visit the same URL at `app.` in case it lives there.

* Add some additional redirects

This add a couple necessary redirects for:
1. paths that should point somewhere else internally
2. paths that should point to external app. URLs

* Move all redirects to next.config.js

and then remove vercel.config because it's empty