Handle access token cookie for shared auth (#183)
* Try refreshUser if an access_token is in cookie

If the authentication server puts an access token in a domain cookie,
then egghead-next should try to use it as a bearer token for requests.

* Label some magic numbers in setSession

* Pass specific values into setSession

Update the two call sites of this function. This opens up a refactor for
the cookied-access token approach to also be able to call setSession.

* Set up session with both auth approaches

Whether the user signs in directly through egghead-next and hits the
`handleAuthentication` flow or signs in over at egghead-rails and hits
the cookie-based access token auth flow
(handleCookieBasedAccessTokenAuthentication), the session will be set
up. This means appropriate local storage and cookie values will be set
and fresh user data will be fetched.

* Remove unneeded changes from getAuthToken

* move stuff around

* Remove catch, flatten then to async/await

Co-authored-by: Joel Hooks <joelhooks@gmail.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>