Add PPP coupon support via commerce XState machine (#917)
* Add commerce machine with support for PPP coupons

* Remove concept of regular coupons from machine

This isn't needed yet, but can easily be added back later if non-PPP
coupons are to be used.

* Update list to show pricing is being recalculated

* Do general clean up of logging and variables

* Move some pricing logic into commerce machine hook

* Tighten up logic for when to display PPP offer

* Debounce quantity changes to reduce UI shift

and to prevent the component from over-fetching the pricing endpoint.

* Move inline assigner to named fun with event type

This is supposed to help with the typing of the `invoke` block.
XState Discord Discussion: https://discord.com/channels/795785288994652170/809564589880639548/895423754790928455

* Show optimistic update of the PPP checkbox

When it is first clicked, show it checked even as prices take a moment
to load.

* Remove old comment

* Match annual plan name to what server returns

* Add proper typing for click handler

* Display skeleton when plan intervals are loading

This is to eliminate UI shift/jank.

* Shift debounced quantity changes for loading UI

This adjustment means that the loading state of the UI is cleaner, not
displaying any NaNs or wrong price-per-seat numbers.

* Add a custom Jest matcher for XState assertions

* Add initial tests for commerceMachine

There are a few pieces of these tests that I couldn't get to pass, so I
left TODOs or skipped them. They can be revisited later.

* Refactor createModel to createMachine

This alleviates the type error on `invoke` that I couldn't otherwise get
rid of.

* Clean up test pricing response

Only included useful values, simplify price IDs

* Remove TODO comments now that the PR is ready

* Add shared type for Coupon from server

* Ensure jest typings are available for matchers