Run pro status check for auth'd user on pricing (#1020)
* Run pro status check for auth'd user on pricing

We are seeing users from time to time that land on the pricing page not
realizing they still have remaining time on their cancelled
subscription, who then try to subscribe anew. They end up on the Stripe
Checkout Session and are able to create a new subscription. We should
instead halt them at the pricing page and direct them to support for any
adjustments they need to make.

* add middleware to redirect pro members to user page from pricing

* filter pricing for pro

* tweak to use access token as logged in status indicator

* viewer isn't always defined

* cast to boolean or switch doesn't work

Co-authored-by: Joel Hooks <joelhooks@gmail.com>