fix: turn CDN off for Sanity client w/ token
I've been seeing a warning in my development logs that you can't set
`useCdn` to true while also using a `token`. I confirmed this in the
docs (https://www.sanity.io/help/js-client-cdn-configuration) as well.
Since this instace of the Sanity Client always has a token, we should
set `useCdn` to false. As the warning put it, we don't want to cache
private data in the CDN.