feat: require one or the other URL to be set
This replaces the required validation on `originalVideoUrl` with a pair
of validations that require one or other of `originalVideoUrl` and
`hslUrl` to be set. We are going to need one or the other, but we don't
always have both.

When a video is first imported via the Course Builder, we only have the
`originalVideoUrl`.

When a Course and its consistent lessons and video resources are
imported from egghead-rails, it is going to have an `hslUrl`, but not an
`originalVideoUrl`.