feat: add explicitly public Last Updated attr
With Sanity, we already have an underlying `_updatedAt` attribute on all
records. That is an internal value that is updated anytime we touch the
record. It's great for internal auditing, but not a value that we want
to display publicly to communicate meaningful updates to the lesson.

Instead, we had originally added an `eggheadRailsUpdatedAt` field as a
destination for syncing the `updated_at` field from egghead-rails. This
naming is off though. What we want this field to represent is a point in
time when the lesson was meaningfully updated which we can display to
the learner. Thus, `displayedUpdatedAt`.

When we import lessons from egghead-rails, we can initially seed that
attribute with whatever is in the egghead-rails `updated_at`. However,
from there, it is a field that can be manually updated by editors as
they see fit.

This commit also renames the `publishedAt` field along similar lines.

The `eggheadRailsCreatedAt` is left as named because that accurately
represents what value that field will contain. It will be synced from
egghead-rails and let you know when the lesson was actually created
instead of when the Sanity record was (much more recently) created.